requirements.txt
eth-brownie>=1.19.0
pandas>=1.4.3
python-dotenv>=0.16.0
dash>=2.6.1
plotly>=5.9.0
